+++ lib/psych/merge/file_analysis.rb
@@ -107,7 +107,7 @@ module Psych
       def comment_augmenter(owners: nil, **options)
         comment_tracker.augment(
           owners: owners || comment_augmenter_default_owners,
-          **options
+          **options,
         )
       end